requirements.txt
gym-derk==0.16.0
numpy
python-telegram-bot
torch
cv2